gateway/requirements.txt
pymongo==4.0.1
python-dotenv==0.19.2
